Atomic CAS loop: close, but no cigar.
s_atomic_cmpswap uses a scalar register pair for the expected value
and the new value, and it returns the old value in the same register
as the new value. This can be used to test if the atomic operation
succeeded (the old value is equal to the expected value), and if not,
to spin by branching back to the s_atomic_cmpswap.

However, there is no concept of registers in IR, let alone of scalar
registers making up a register pair. To get individual registers,
say, s4 and s5, out of s[4:5], you have to do a truncation to 32 bits
of the Value* representing s[4:5] to get s4, and a 32-bits right
shift and a truncation for s5. This looks inefficient in the IR,
but in the final ISA, no truncation or shifting is done.

The problem with this approach is that we now have to split the inline
asm in two parts, with the truncation and shifting using IR functionality
in between. The second part of the inline assembly has the comparison
operation and the branch, with the branch target in the first part of the
inline assembly. But there's no guarantee that the two inline asm
parts end up consecutively in the final ISA, and in fact, for a small
test program, it didn't for one out of 8 cases, in which there were to
register copy operations in between the two inline assembly parts. This
messes up the branching between the two inline assembly parts, and even
if the branching could be fixed (and it can), we now have additional
instructions in our CAS loop that are executed repeatedly if we loop indeed.
This may affect correctness of the program.

The fix is to not compare the individual registers with each other,
but to compare the register pair with another register pair. For instance,
instead of
  0 and 0 -> success, don't loop
  1 and 0 -> fail, loop
we compare
  0:0 and 1:0 -> success, don't loop
  1:0 and 1:0 -> fail, loop

With this approach, we don't need the IR functionality for extracting
32-bits values out of 64-bits values, which means that we no longer have
to split the inline asm. As a result, the compiler will no longer
move other instructions inside our CAS loop.