Closer, still no cigar.
Leaving this commit in for educational puprposes, and to remind myself
what didn't quite work, and why.

Setting constrainst for inline asm with loops is tricky. Final ISA is correct
for all but the last two executions of the CAS loop, where the register
assignment went wrong. Optimizer decided to reuse read-constrained registers
that were no longer needed after the CAS loop, resulting in

  s_cmp_eq_u64 s[2:3], s[2:3]

which always evaluates to true. Hmmm...