Fix issues with instrumenting code with multiple kernels.
Value* TtraceCounter was declared outside loop over kernels, causing
linking errors, since it was not a global variable, but used in multiple
kernels. Moved the declaration inside the loop, and initialized it
with a counter that is updated across loop iterations.

Moved printing of statistics of instrumented LDS instructions
outside the loop as well.